Add context menu to open a url from a short log.
The url is built from a regex expression found in the short log, and a prefix from Settings. Added a new "External link" tab in Settings to enable this feature, and to specify the regex and the url prefix.
This can be very useful when Short Log includes some kind of reference to a bug number, and that bug can be accessed through a url.